requirement.txt
Flask==3.0.3
pymongo[srv]==3.12
python-dotenv==1.0.1
